Motion Tracking Software
I have done most of the edits for the queue handling but I'm not 100% done yet. I'm using a WatchService interface to watch the queue directory for creation or modification changes. I think I might have to add another loop or change the loop a little to accommodate the way the WatchService works.